+++ hutool-core/src/main/java/cn/hutool/core/lang/caller/StackTraceCaller.java
@@ -16,7 +16,7 @@ public class StackTraceCaller implements Caller, Serializable {
 	@Override
 	public Class<?> getCaller() {
 		final StackTraceElement[] stackTrace = Thread.currentThread().getStackTrace();
-		if (null == stackTrace || (OFFSET + 1) >= stackTrace.length) {
+		if (OFFSET + 1 >= stackTrace.length) {
 			return null;
 		}
 		final String className = stackTrace[OFFSET + 1].getClassName();
@@ -30,7 +30,7 @@ public class StackTraceCaller implements Caller, Serializable {
 	@Override
 	public Class<?> getCallerCaller() {
 		final StackTraceElement[] stackTrace = Thread.currentThread().getStackTrace();
-		if (null == stackTrace || (OFFSET + 2) >= stackTrace.length) {
+		if (OFFSET + 2 >= stackTrace.length) {
 			return null;
 		}
 		final String className = stackTrace[OFFSET + 2].getClassName();
@@ -44,7 +44,7 @@ public class StackTraceCaller implements Caller, Serializable {
 	@Override
 	public Class<?> getCaller(int depth) {
 		final StackTraceElement[] stackTrace = Thread.currentThread().getStackTrace();
-		if (null == stackTrace || (OFFSET + depth) >= stackTrace.length) {
+		if (OFFSET + depth >= stackTrace.length) {
 			return null;
 		}
 		final String className = stackTrace[OFFSET + depth].getClassName();
@@ -58,11 +58,9 @@ public class StackTraceCaller implements Caller, Serializable {
 	@Override
 	public boolean isCalledBy(Class<?> clazz) {
 		final StackTraceElement[] stackTrace = Thread.currentThread().getStackTrace();
-		if(null != stackTrace) {
-			for (final StackTraceElement element : stackTrace) {
-				if (element.getClassName().equals(clazz.getName())) {
-					return true;
-				}
+		for (final StackTraceElement element : stackTrace) {
+			if (element.getClassName().equals(clazz.getName())) {
+				return true;
 			}
 		}
 		return false;