+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * 【core   】     BooleanUtil中重载歧义修正，修改了包装参数的方法名（issue#I1BSK8@Gitee）
 * 【core   】     XmlUtil增加xmlToBean和beanToXml方法
 * 【db     】     设置全局忽略大小写DbUtil.setCaseInsensitiveGlobal(true)（issue#784@Github）
+* 【core   】     增加CallerUtil.getCallerMethodName方法
+* 【core   】     Tree增加getParent方法，可以获取父节点
 
 ### Bug修复
 * 【core   】     修复TypeUtil无法获取泛型接口的泛型参数问题（issue#I1BRFI@Gitee）